# portable_simd version of Avg (4bpp)