Implements a RGBA version of the Avg filter with `portable_simd` intrinsics. 

| CPU  | Baseline | Result | Speedup |
| ---  | --- | --- | --- |
| Arm  Cortex A520 | 415.9 MiB/s | 707.4 MiB/s | 70.08% |
| Arm Cortex X4 | 2053.9 MiB/s | 2334.9 MiB/s | 13.68% |
| Apple Silicon M2 | 2053.9 MiB/s | 2173.5 MiB/s | 3.62% |
| AMD EPYC 7B13 | 2425.8 MiB/s | 2150.8 MiB/s | -11.34% |

Marked as draft until #632 is completed.