@Sentimentron 2025-09-09
AI disclosure: I wrote a original sliding-window `portable_simd` implementation of the Paeth filter (3bpp) and optimized it for best performance on the Cortex A520. I then used the Gemini family of LLMs provided by my employer to automatically adapt this code to the Avg filter from a written description, then optimize it to achieve the best possible code-generation and performance across all other micro-architectures in simulation. This PR is derived from that output, but includes documentation and other cleanups.